python-worker/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.0
sentence-transformers==3.3.1
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.4
supabase==2.3.0
tenacity==8.2.3
httpx==0.25.2
